pyproject.toml
[project]
name = "chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.12.13",
    "aiosqlite>=0.21.0",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.5",
    "langchain-groq>=0.3.4",
    "langgraph>=0.4.10",
    "python-dotenv>=1.1.1",
    "teams-ai>=1.6.0,<2.0.0",
]
